split notes src/dst into notes.lua #1
